fix: [F-012] tickets-on-merge.yml fixing execution the final state after merge
