+++ .github/workflows/tickets-on-merge.yml
@@ -4,6 +4,16 @@ on:
   pull_request:
     types: [closed]
     branches: [master]
+  workflow_dispatch:
+    inputs:
+      mode:
+        description: "Sync mode"
+        required: true
+        default: "changed"
+        type: choice
+        options:
+          - changed
+          - full
 
 permissions:
   contents: write
@@ -11,7 +21,7 @@ permissions:
 
 jobs:
   on-merge:
-    if: github.event.pull_request.merged == true
+    if: github.event.pull_request.merged == true || github.event_name == 'workflow_dispatch'
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
@@ -25,6 +35,8 @@ jobs:
 
       - name: Close issues + mark files as Merged + sync Project v2
         env:
+          TICKETS_MODE: merge
+          SYNC_MODE: ${{ github.event.inputs.mode || 'changed' }}
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
           PROJECT_TOKEN: ${{ secrets.PROJECT_TOKEN }}
 
@@ -37,4 +49,4 @@ jobs:
 
           PROJECT_FIELD_STATUS: Status
           PROJECT_FIELD_AREA: Area
-        run: node .github/scripts/tickets-on-merge.js
+        run: node .github/scripts/tickets.js
